backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
Pillow==10.2.0
python-decouple==3.8

